+++ test/utilTest.js
@@ -20,7 +20,7 @@ describe('util', function() {
       ],
       'flash': ['application/x-shockwave-flash'],
       'favicon': ['image/x-icon', 'image/vnd.microsoft.icon'],
-      'others': ['application/my-own-type']
+      'other': ['application/my-own-type']
     };
 
     forEach(typeToMimeMapping, function(mimes, type) {